+++ lib/DevApp/src/main/java/dev/utils/app/AppUtils.java
@@ -27,6 +27,7 @@ import android.media.AudioManager;
 import android.net.ConnectivityManager;
 import android.net.Uri;
 import android.net.wifi.WifiManager;
+import android.os.Build;
 import android.os.PowerManager;
 import android.os.Vibrator;
 import android.os.storage.StorageManager;
@@ -433,7 +434,7 @@ public final class AppUtils {
      * 获取 APP versionCode
      * @return APP versionCode
      */
-    public static int getAppVersionCode() {
+    public static long getAppVersionCode() {
         return getAppVersionCode(getPackageName());
     }
 
@@ -442,11 +443,15 @@ public final class AppUtils {
      * @param packageName 应用包名
      * @return APP versionCode
      */
-    public static int getAppVersionCode(final String packageName) {
+    public static long getAppVersionCode(final String packageName) {
         if (StringUtils.isSpace(packageName)) return -1;
         try {
             PackageInfo packageInfo = getPackageInfo(packageName, PackageManager.GET_SIGNATURES);
-            return packageInfo == null ? -1 : packageInfo.versionCode;
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                return packageInfo.getLongVersionCode();
+            } else {
+                return packageInfo.versionCode;
+            }
         } catch (Exception e) {
             LogPrintUtils.eTag(TAG, e, "getAppVersionCode");
             return -1;
@@ -712,7 +717,7 @@ public final class AppUtils {
         if (StringUtils.isSpace(packageName)) return false;
         try {
             ApplicationInfo appInfo = getApplicationInfo(packageName, PackageManager.GET_UNINSTALLED_PACKAGES);
-            return true;
+            return appInfo != null;
         } catch (Exception e) { // 未安装, 则会抛出异常
             LogPrintUtils.eTag(TAG, e, "isInstalledApp");
             return false;