+++ lib/DevApp/src/main/java/dev/utils/app/DeviceUtils.java
@@ -5,7 +5,6 @@ import android.content.Intent;
 import android.net.wifi.WifiInfo;
 import android.net.wifi.WifiManager;
 import android.os.Build;
-import android.os.PowerManager;
 import android.provider.Settings;
 import android.text.TextUtils;
 
@@ -674,8 +673,7 @@ public final class DeviceUtils {
      */
     public static boolean reboot(final String reason) {
         try {
-            PowerManager mPowerManager = AppUtils.getPowerManager();
-            mPowerManager.reboot(reason);
+            AppUtils.getPowerManager().reboot(reason);
             return true;
         } catch (Exception e) {
             LogPrintUtils.eTag(TAG, e, "reboot");